review deletion

+++ models/listingmodel.js
@@ -40,7 +40,10 @@ const listingSchema = new Schema({
 
 
 listingSchema.post("findOneAndDelete", async (listing) => {
-  await reviews.deleteMany({ _id: { $in: listing.review } });
+  if(listing){
+    await reviews.deleteMany({ _id: { $in: listing.review } });
+  }
+  
 });
 
 const Listing = mongoose.model("Listing", listingSchema);
